Refactor client
Moving the parser into transport
