+++ lua/conjure/sponsors.lua
@@ -1,2 +1,2 @@
 local _2afile_2a = "fnl/conjure/sponsors.fnl"
-return {"Aljendro", "alloy-d", "anthony-khong", "beanpuppy", "BerkeleyTrue", "campbellr", "daveyarwood", "davidmh", "dharrigan", "djwhitt", "Fedreg", "frenchy64", "fuadsaud", "harrygallagher4", "jcpsantiago", "jkrasnay", "ketansrivastav", "kevinkrouse", "lucasdf", "lucaslollobrigida", "mamapitufo", "martinklepsch", "pyrmont", "qnkhuat", "rafaeldelboni", "rbatista", "rgm", "stelcodes", "thiru", "TimoKramer", "zhming0"}
+return {"Aljendro", "alloy-d", "anthony-khong", "beanpuppy", "BerkeleyTrue", "campbellr", "daveyarwood", "davidmh", "dharrigan", "djwhitt", "Fedreg", "frenchy64", "fuadsaud", "harrygallagher4", "jcpsantiago", "jkrasnay", "kevinkrouse", "lucasdf", "lucaslollobrigida", "mamapitufo", "martinklepsch", "pyrmont", "qnkhuat", "rafaeldelboni", "rbatista", "rgm", "stelcodes", "thiru", "TimoKramer", "zhming0"}
